refactoring and LICENSE updated
