+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright 2021-2023 Giulio Ermanno Pibiri
+Copyright 2021-2024 Giulio Ermanno Pibiri, Oleksandr Kulkov, and COMBINE Lab
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
