+++ external/gz/zip_stream.cpp
