+++ external/gz/zip_stream_fwd.hpp
