+++ include/builder/parse_file.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "include/gz/zip_stream.hpp"
+#include "external/gz/zip_stream.hpp"
 
 namespace sshash {
