+++ tools/common.hpp
