+++ tools/permute.cpp
