+++ tools/query.cpp
