+++ tools/sshash.cpp
